feat(projects)#1285 add types and add comments
