+++ functions/src/index.ts
@@ -1,25 +1,34 @@
+// Third party modules
 import * as functions from 'firebase-functions';
 import { CallableContext } from 'firebase-functions/lib/providers/https';
-import { getRepositoryInfo, RepositoryInfoInput } from './repository/info';
 
-import { getUserEvents, EventsInput } from './user/events';
-import { getUserRepos, ReposInput } from './user/repos';
+// Dashboard hub firebase functions models/mappers
+import { DocumentSnapshot } from './client/firebase-admin';
 
-import { onDeleteProjectRepositories } from './project/delete-project';
-import { onUpdateProjectRepositories } from './project/update-repositories';
+// Dashboard repositories
 import { onCreateRepository } from './repository/create-repository';
+import { getRepositoryInfo, RepositoryInfoInput } from './repository/info';
 import { onUpdateDeleteRepository } from './repository/update-delete-repository';
 
+// Dashboard users
+import { getUserEvents, EventsInput } from './user/events';
+import { getUserRepos, ReposInput } from './user/repos';
+import { onUpdateUserStats } from './user/stats';
+
+// Dashboard projects
+import { onDeleteProjectRepositories } from './project/delete-project';
+import { onUpdateProjectRepositories } from './project/update-repositories';
 
 declare type HttpsFunction = functions.HttpsFunction;
-
-export * from './user/stats';
+declare type CloudFunction<T> = functions.CloudFunction<T>;
+declare type Change<T> = functions.Change<T>;
 
 export const findAllUserRepositories: HttpsFunction = functions.https.onCall((input: ReposInput, context: CallableContext) => getUserRepos(input.token, context.auth.uid));
 export const findAllUserEvents: HttpsFunction = functions.https.onCall((input: EventsInput, context: CallableContext) => getUserEvents(input.token, context.auth.uid, input.username));
 export const findRepositoryInfo: HttpsFunction = functions.https.onCall((input: RepositoryInfoInput, context: CallableContext) => getRepositoryInfo(input.token, input.fullName));
 
-export const deleteProjectRepositories: any = onDeleteProjectRepositories;
-export const updateProjectRepositories: any = onUpdateProjectRepositories;
-export const updateDeleteRepository: any = onUpdateDeleteRepository;
-export const createRepository: any = onCreateRepository;
+export const deleteProjectRepositories: CloudFunction<DocumentSnapshot> = onDeleteProjectRepositories;
+export const updateProjectRepositories: CloudFunction<DocumentSnapshot> = onUpdateProjectRepositories;
+export const updateDeleteRepository: CloudFunction<Change<DocumentSnapshot>> = onUpdateDeleteRepository;
+export const createRepository: CloudFunction<DocumentSnapshot> = onCreateRepository;
+export const updateUserStats: CloudFunction<DocumentSnapshot> = onUpdateUserStats;
