+++ functions/src/repository/update-delete-repository.ts
@@ -1,13 +1,8 @@
 // Third party modules
-import * as admin from 'firebase-admin';
 import { firestore, Change, CloudFunction, EventContext } from 'firebase-functions';
 
 // Dashboard hub firebase functions models/mappers
-import { FirebaseAdmin } from './../client/firebase-admin';
-
-
-export declare type DocumentSnapshot = admin.firestore.DocumentSnapshot;
-export declare type DocumentData = admin.firestore.DocumentData;
+import { DocumentData, DocumentSnapshot, FirebaseAdmin } from './../client/firebase-admin';
 
 export const onUpdateDeleteRepository: CloudFunction<Change<DocumentSnapshot>> = firestore
   .document('repositories/{repositoryUid}')
